Extract the method stub context record type to its own file.
